# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EvaluateActionRequest {
  /** The scenario text to evaluate */
  scenario: string;
  /** Additional context for the scenario */
  context?: string;
}

export interface EthicalFrameworkScores {
  /**
   * @minimum 0
   * @maximum 100
   */
  utilitarian: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  deontological: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  virtue: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  care: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  context: number;
}

export type EthicalEvaluationDecision =
  (typeof EthicalEvaluationDecision)[keyof typeof EthicalEvaluationDecision];

export const EthicalEvaluationDecision = {
  APPROVED: "APPROVED",
  CONDITIONAL: "CONDITIONAL",
  FLAGGED: "FLAGGED",
  BLOCKED: "BLOCKED",
} as const;

export type EthicalEvaluationOverrideDecision =
  | (typeof EthicalEvaluationOverrideDecision)[keyof typeof EthicalEvaluationOverrideDecision]
  | null;

export const EthicalEvaluationOverrideDecision = {
  APPROVED: "APPROVED",
  CONDITIONAL: "CONDITIONAL",
  FLAGGED: "FLAGGED",
  BLOCKED: "BLOCKED",
} as const;

export interface EthicalEvaluation {
  id: number;
  scenario: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  ethicalScore: number;
  decision: EthicalEvaluationDecision;
  explanation: string;
  alternatives: string[];
  frameworkScores: EthicalFrameworkScores;
  timestamp: string;
  isOverridden: boolean;
  overrideDecision?: EthicalEvaluationOverrideDecision;
}

/**
 * Optional category for scenario generation
 */
export type GenerateScenarioRequestCategory =
  (typeof GenerateScenarioRequestCategory)[keyof typeof GenerateScenarioRequestCategory];

export const GenerateScenarioRequestCategory = {
  "trolley-problem": "trolley-problem",
  "resource-allocation": "resource-allocation",
  privacy: "privacy",
  autonomy: "autonomy",
  safety: "safety",
  environmental: "environmental",
} as const;

export interface GenerateScenarioRequest {
  /** Optional category for scenario generation */
  category?: GenerateScenarioRequestCategory;
}

export interface GeneratedScenario {
  scenario: string;
  category: string;
}

export interface HistoryResponse {
  decisions: EthicalEvaluation[];
  total: number;
  offset: number;
  limit: number;
}

export type FeedbackRequestUserFeedback =
  (typeof FeedbackRequestUserFeedback)[keyof typeof FeedbackRequestUserFeedback];

export const FeedbackRequestUserFeedback = {
  approve: "approve",
  reject: "reject",
} as const;

export interface FeedbackRequest {
  decisionId: number;
  userFeedback: FeedbackRequestUserFeedback;
  comment?: string;
}

export interface FeedbackResponse {
  id: number;
  decisionId: number;
  userFeedback: string;
  timestamp: string;
}

export type SystemStatusStatus =
  (typeof SystemStatusStatus)[keyof typeof SystemStatusStatus];

export const SystemStatusStatus = {
  online: "online",
  degraded: "degraded",
  offline: "offline",
} as const;

export interface SystemStatus {
  status: SystemStatusStatus;
  totalDecisions: number;
  approvedCount: number;
  conditionalCount: number;
  flaggedCount: number;
  blockedCount: number;
  averageScore: number;
  llmAvailable: boolean;
  uptime: number;
}

export interface EthicalWeights {
  /**
   * @minimum 0
   * @maximum 1
   */
  utilitarian: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  deontological: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  virtue: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  care: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  context: number;
}

export type OverrideRequestOverrideDecision =
  (typeof OverrideRequestOverrideDecision)[keyof typeof OverrideRequestOverrideDecision];

export const OverrideRequestOverrideDecision = {
  APPROVED: "APPROVED",
  CONDITIONAL: "CONDITIONAL",
  FLAGGED: "FLAGGED",
  BLOCKED: "BLOCKED",
} as const;

export interface OverrideRequest {
  decisionId: number;
  overrideDecision: OverrideRequestOverrideDecision;
  reason: string;
}

export type GetHistoryParams = {
  limit?: number;
  offset?: number;
  decision?: GetHistoryDecision;
};

export type GetHistoryDecision =
  (typeof GetHistoryDecision)[keyof typeof GetHistoryDecision];

export const GetHistoryDecision = {
  APPROVED: "APPROVED",
  CONDITIONAL: "CONDITIONAL",
  FLAGGED: "FLAGGED",
  BLOCKED: "BLOCKED",
} as const;