# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AECE - Autonomous Ethical Cognition Engine API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ethics
    description: Ethical evaluation operations
  - name: scenarios
    description: Scenario generation
  - name: history
    description: Decision history
  - name: feedback
    description: User feedback
  - name: governance
    description: Governance and policy control
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /evaluate-action:
    post:
      operationId: evaluateAction
      tags: [ethics]
      summary: Evaluate an ethical action
      description: Evaluate a scenario for ethical compliance using multi-framework analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateActionRequest"
      responses:
        "200":
          description: Ethical evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EthicalEvaluation"

  /generate-scenario:
    post:
      operationId: generateScenario
      tags: [scenarios]
      summary: Generate a random ethical scenario
      description: Auto-generate a robot ethical scenario for testing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScenarioRequest"
      responses:
        "200":
          description: Generated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedScenario"

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get decision history
      description: Return all stored ethical decisions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: decision
          in: query
          schema:
            type: string
            enum: [APPROVED, CONDITIONAL, FLAGGED, BLOCKED]
      responses:
        "200":
          description: List of decisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"

  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback on a decision
      description: Store user feedback (approve/reject) on an ethical decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"

  /system-status:
    get:
      operationId: getSystemStatus
      tags: [health]
      summary: Get system status
      description: Returns overall system health and statistics
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStatus"

  /weights:
    get:
      operationId: getWeights
      tags: [governance]
      summary: Get current ethical framework weights
      responses:
        "200":
          description: Current weights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EthicalWeights"
    put:
      operationId: updateWeights
      tags: [governance]
      summary: Update ethical framework weights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EthicalWeights"
      responses:
        "200":
          description: Updated weights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EthicalWeights"

  /override:
    post:
      operationId: overrideDecision
      tags: [governance]
      summary: Manually override a decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideRequest"
      responses:
        "200":
          description: Override applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EthicalEvaluation"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    EvaluateActionRequest:
      type: object
      properties:
        scenario:
          type: string
          description: The scenario text to evaluate
        context:
          type: string
          description: Additional context for the scenario
      required:
        - scenario

    EthicalFrameworkScores:
      type: object
      properties:
        utilitarian:
          type: number
          minimum: 0
          maximum: 100
        deontological:
          type: number
          minimum: 0
          maximum: 100
        virtue:
          type: number
          minimum: 0
          maximum: 100
        care:
          type: number
          minimum: 0
          maximum: 100
        context:
          type: number
          minimum: 0
          maximum: 100
      required:
        - utilitarian
        - deontological
        - virtue
        - care
        - context

    EthicalEvaluation:
      type: object
      properties:
        id:
          type: integer
        scenario:
          type: string
        ethicalScore:
          type: number
          minimum: 0
          maximum: 100
        decision:
          type: string
          enum: [APPROVED, CONDITIONAL, FLAGGED, BLOCKED]
        explanation:
          type: string
        alternatives:
          type: array
          items:
            type: string
        frameworkScores:
          $ref: "#/components/schemas/EthicalFrameworkScores"
        timestamp:
          type: string
          format: date-time
        isOverridden:
          type: boolean
        overrideDecision:
          type: string
          nullable: true
          enum: [APPROVED, CONDITIONAL, FLAGGED, BLOCKED, null]
      required:
        - id
        - scenario
        - ethicalScore
        - decision
        - explanation
        - alternatives
        - frameworkScores
        - timestamp
        - isOverridden

    GenerateScenarioRequest:
      type: object
      properties:
        category:
          type: string
          description: Optional category for scenario generation
          enum: [trolley-problem, resource-allocation, privacy, autonomy, safety, environmental]

    GeneratedScenario:
      type: object
      properties:
        scenario:
          type: string
        category:
          type: string
      required:
        - scenario
        - category

    HistoryResponse:
      type: object
      properties:
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/EthicalEvaluation"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - decisions
        - total
        - offset
        - limit

    FeedbackRequest:
      type: object
      properties:
        decisionId:
          type: integer
        userFeedback:
          type: string
          enum: [approve, reject]
        comment:
          type: string
      required:
        - decisionId
        - userFeedback

    FeedbackResponse:
      type: object
      properties:
        id:
          type: integer
        decisionId:
          type: integer
        userFeedback:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - decisionId
        - userFeedback
        - timestamp

    SystemStatus:
      type: object
      properties:
        status:
          type: string
          enum: [online, degraded, offline]
        totalDecisions:
          type: integer
        approvedCount:
          type: integer
        conditionalCount:
          type: integer
        flaggedCount:
          type: integer
        blockedCount:
          type: integer
        averageScore:
          type: number
        llmAvailable:
          type: boolean
        uptime:
          type: number
      required:
        - status
        - totalDecisions
        - approvedCount
        - conditionalCount
        - flaggedCount
        - blockedCount
        - averageScore
        - llmAvailable
        - uptime

    EthicalWeights:
      type: object
      properties:
        utilitarian:
          type: number
          minimum: 0
          maximum: 1
        deontological:
          type: number
          minimum: 0
          maximum: 1
        virtue:
          type: number
          minimum: 0
          maximum: 1
        care:
          type: number
          minimum: 0
          maximum: 1
        context:
          type: number
          minimum: 0
          maximum: 1
      required:
        - utilitarian
        - deontological
        - virtue
        - care
        - context

    OverrideRequest:
      type: object
      properties:
        decisionId:
          type: integer
        overrideDecision:
          type: string
          enum: [APPROVED, CONDITIONAL, FLAGGED, BLOCKED]
        reason:
          type: string
      required:
        - decisionId
        - overrideDecision
        - reason