# lib/api-zod/src/generated/types/ethicalEvaluation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { EthicalEvaluationDecision } from "./ethicalEvaluationDecision";
import type { EthicalEvaluationOverrideDecision } from "./ethicalEvaluationOverrideDecision";
import type { EthicalFrameworkScores } from "./ethicalFrameworkScores";

export interface EthicalEvaluation {
  id: number;
  scenario: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  ethicalScore: number;
  decision: EthicalEvaluationDecision;
  explanation: string;
  alternatives: string[];
  frameworkScores: EthicalFrameworkScores;
  timestamp: Date;
  isOverridden: boolean;
  overrideDecision?: EthicalEvaluationOverrideDecision;
}