# lib/api-zod/src/generated/types/ethicalWeights.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface EthicalWeights {
  /**
   * @minimum 0
   * @maximum 1
   */
  utilitarian: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  deontological: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  virtue: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  care: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  context: number;
}