# lib/api-zod/src/generated/types/generateScenarioRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateScenarioRequestCategory } from "./generateScenarioRequestCategory";

export interface GenerateScenarioRequest {
  /** Optional category for scenario generation */
  category?: GenerateScenarioRequestCategory;
}