# lib/api-zod/src/generated/types/generatedScenario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedScenario {
  scenario: string;
  category: string;
}