# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AECE - Autonomous Ethical Cognition Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { EthicalEvaluation } from "./ethicalEvaluation";

export interface HistoryResponse {
  decisions: EthicalEvaluation[];
  total: number;
  offset: number;
  limit: number;
}